+++ TeamsEmbediOSGettingStarted/ACSViewController.swift
@@ -1,0 +1,12 @@
+//----------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+//
+// Name: ACSViewController.swift
+//----------------------------------------------------------------
+
+import Foundation
+
+import UIKit
+
+class ACSViewController: UIViewController {
+}
