+++ TeamsEmbediOSGettingStarted/ViewController.swift
@@ -7,25 +7,18 @@
 import UIKit
 
 class ViewController: UIViewController {
-
+    
     private let acsToken = "<ACS_TOKEN>"
-    
-    private var teamsSdkController: TeamsEmbedSdkController?
     private var acsSdkController: AcsSdkController?
-    
     public let statusLabel = UILabel(frame: CGRect(x: 100, y: 100, width: 200, height: 50))
-    
-    let teamsSdkLabel = UILabel(frame: CGRect(x: 100, y: 100, width: 200, height: 50))
     let acsSdkLabel = UILabel(frame: CGRect(x: 100, y: 100, width: 200, height: 50))
-    
-    let joinMeetingButton = Button(text: "Join Meeting")
-    let joinGroupCallButton = Button(text: "Join Group Call")
-    let endMeetingButton = Button(text: "Hang up and Stop Teams")
-    
     let joinAcsCallButton = Button(text: "Call 8:echo123")
     let endAcsCallButton = Button(text: "End Call")
     let stopAcsCallButton = Button(text: "Stop ACS")
     
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -38,42 +31,13 @@ class ViewController: UIViewController {
         statusLabel.centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
         statusLabel.topAnchor.constraint(equalTo: view.topAnchor, constant: 80).isActive = true
         
-        teamsSdkLabel.textColor = .systemPurple
-        teamsSdkLabel.text = "----- Teams Embed SDK -----"
-        
-        teamsSdkLabel.translatesAutoresizingMaskIntoConstraints = false
-        self.view.addSubview(teamsSdkLabel)
-        teamsSdkLabel.centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
-        teamsSdkLabel.topAnchor.constraint(equalTo: statusLabel.bottomAnchor, constant: 80).isActive = true
-        
-        joinMeetingButton.addTarget(self, action: #selector(joinMeetingTapped), for: .touchUpInside)
-        
-        joinMeetingButton.translatesAutoresizingMaskIntoConstraints = false
-        self.view.addSubview(joinMeetingButton)
-        joinMeetingButton.centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
-        joinMeetingButton.topAnchor.constraint(equalTo: teamsSdkLabel.bottomAnchor, constant: 10).isActive = true
-                        
-        joinGroupCallButton.addTarget(self, action: #selector(joinGroupCallTapped), for: .touchUpInside)
-        
-        joinGroupCallButton.translatesAutoresizingMaskIntoConstraints = false
-        self.view.addSubview(joinGroupCallButton)
-        joinGroupCallButton.centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
-        joinGroupCallButton.topAnchor.constraint(equalTo: joinMeetingButton.bottomAnchor, constant: 20).isActive = true
-
-        endMeetingButton.addTarget(self, action: #selector(endMeetingTapped), for: .touchUpInside)
-        
-        endMeetingButton.translatesAutoresizingMaskIntoConstraints = false
-        self.view.addSubview(endMeetingButton)
-        endMeetingButton.centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
-        endMeetingButton.topAnchor.constraint(equalTo: joinGroupCallButton.bottomAnchor, constant: 20).isActive = true
-        
         acsSdkLabel.textColor = .systemGreen
         acsSdkLabel.text = "----- ACS Calling SDK -----"
         
         acsSdkLabel.translatesAutoresizingMaskIntoConstraints = false
         self.view.addSubview(acsSdkLabel)
         acsSdkLabel.centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
-        acsSdkLabel.topAnchor.constraint(equalTo: endMeetingButton.bottomAnchor, constant: 50).isActive = true
+        acsSdkLabel.topAnchor.constraint(equalTo: statusLabel.bottomAnchor, constant: 80).isActive = true
         
         joinAcsCallButton.addTarget(self, action: #selector(joinAcsCallTapped), for: .touchUpInside)
         
@@ -96,32 +60,19 @@ class ViewController: UIViewController {
         stopAcsCallButton.centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
         stopAcsCallButton.topAnchor.constraint(equalTo: endAcsCallButton.bottomAnchor, constant: 20).isActive = true
         
-        self.teamsSdkController = TeamsEmbedSdkController(with: self.acsToken, viewController: self)
-        self.acsSdkController = AcsSdkController(with: self.acsToken, viewController: self)
         
+        self.acsSdkController = AcsSdkController(with: self.acsToken, viewController: self)
     }
     
-    @IBAction func joinMeetingTapped(_ sender: UIButton) {
-        self.teamsSdkController?.joinMeeting()
-    }
-    
-    @IBAction func joinGroupCallTapped(_ sender: UIButton) {
-        self.teamsSdkController?.joinGroupCall()
-    }
-    
-    @IBAction func endMeetingTapped(_ sender: UIButton) {
-        self.teamsSdkController?.endMeeting()
-    }
-    
-    @IBAction func joinAcsCallTapped(_ sender: UIButton) {
+    @objc func joinAcsCallTapped() {
         self.acsSdkController?.joinAcsCall()
     }
     
-    @IBAction func endAcsCallTapped(_ sender: UIButton) {
+    @objc func endAcsCallTapped() {
         self.acsSdkController?.endAcsCall()
     }
     
-    @IBAction func stopAcsCallTapped(_ sender: UIButton) {
+    @objc func stopAcsCallTapped() {
         self.acsSdkController?.stopAcs()
     }
        
@@ -135,17 +86,5 @@ class ViewController: UIViewController {
         joinAcsCallButton.disable()
         endAcsCallButton.disable()
         stopAcsCallButton.disable()
-    }
-    
-    public func enableTeamsSdkButtons() {
-        joinMeetingButton.enable()
-        joinGroupCallButton.enable()
-        endMeetingButton.enable()
-    }
-    
-    public func disableTeamsSdkButtons() {
-        joinMeetingButton.disable()
-        joinGroupCallButton.disable()
-        endMeetingButton.disable()
     }
 }
